requirements.txt
fastapi
uvicorn
python-dotenv
vercel-blob
python-multipart
pydantic
Werkzeug
